fix delete

+++ CarDealerShip/Inventory.xaml.cs
@@ -48,21 +48,35 @@ namespace CarDealerShip
             {
                 inventory selectedInventory = DGridInventory.SelectedItem as inventory;
 
-                try
+                MessageBoxResult result = MessageBox.Show("Вы уверены, что хотите удалить эту запись?",
+                                                           "Удаление записи",
+                                                           MessageBoxButton.YesNo,
+                                                           MessageBoxImage.Question);
+
+                if (result == MessageBoxResult.Yes)
                 {
-                    db.inventories.Attach(selectedInventory);
-                    db.inventories.Remove(selectedInventory);
+                    try
+                    {
+                        // Прикрепляем выбранную запись к контексту данных и удаляем ее
+                        db.inventories.Attach(selectedInventory);
+                        db.inventories.Remove(selectedInventory);
 
-                    db.SaveChanges();
+                        db.SaveChanges();
 
-                    MessageBox.Show("Данные удалены");
+                        MessageBox.Show("Данные удалены");
 
-                    LoadInventoryData();
+                        LoadInventoryData();
+                    }
+                    catch (Exception ex)
+                    {
+
+                        MessageBox.Show($"Ошибка при удалении данных: {ex.Message}");
+                    }
                 }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message);
-                }
+            }
+            else
+            {
+                MessageBox.Show("Выберите запись для удаления", "Ошибка", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
